requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.6
django-bootstrap5==24.2
django-crispy-forms==2.1
mysqlclient==2.2.4
sqlparse==0.5.0
tzdata==2024.1
